Final fix threads inactive and qsfs process exit issue in fuse background mode. Description: If run fuse in background mode (without -f), the worker threads of thread pool will get killed after enter fuse event loop, and af ter umount mountpoint the qsfs process still alive. Root Cause: For the thread issue, as libfuse FAQ says Threads should startd from fuse init() method, threads stared before fuse_main() will exit when process goes into the background. For the process issue, qsfs probably need to cleanup drive in fuse destory. Resolution: 1. For threads issue: Warpper the ThreadPool::Initialize to start worker thread, and add ThreadPoolInitializer to initialize each registered thread pool instance, And register all these thread pool instance which need to be start after fuse_main to the ThreadPoolInitializer. Call ThreadPoolInitializer::DoInitialize in fuse init() method; 2. For the process issue: 1) pass drive handle (pointer) to fuse_main which will be suppliced in the context during fuse init(). 2) In init() method, return fuse_context->private_data which will be supplied as parameter to destory(); 3) in destroy() call drive::CleanUp to clean up filesystem on exit. Notes: for process issue, the step 3) could be neglect, this probably because that Drive is a static memory which managed by heap.

+++ src/filesystem/Drive.cpp
@@ -110,6 +110,7 @@ Drive &Drive::Instance() {
 // --------------------------------------------------------------------------
 Drive::Drive()
     : m_mountable(true),
+      m_cleanup(false),
       m_client(ClientFactory::Instance().MakeClient()),
       m_transferManager(std::move(
           TransferManagerFactory::Create(TransferManagerConfigure()))),
@@ -124,24 +125,32 @@ Drive::Drive()
 }
 
 // --------------------------------------------------------------------------
-Drive::~Drive() {
-  // abort unfinished multipart uploads
-  if (!m_unfinishedMultipartUploadHandles.empty()) {
-    for (auto &fileToHandle : m_unfinishedMultipartUploadHandles) {
-      m_transferManager->AbortMultipartUpload(fileToHandle.second);
-    }
-  }
-  // remove temp folder if existing
-  auto tmpfolder = GetCacheTemporaryDirectory();
-  if (FileExists(tmpfolder, false) && IsDirectory(tmpfolder, true)) {  // log on
-    DeleteFilesInDirectory(tmpfolder, true);  // delete folder itself
-  }
+Drive::~Drive() { CleanUp(); }
 
-  m_client.reset();
-  m_transferManager.reset();
-  m_cache.reset();
-  m_directoryTree.reset();
-  m_unfinishedMultipartUploadHandles.clear();
+// --------------------------------------------------------------------------
+void Drive::CleanUp() {
+  if (!m_cleanup) {
+    // abort unfinished multipart uploads
+    if (!m_unfinishedMultipartUploadHandles.empty()) {
+      for (auto &fileToHandle : m_unfinishedMultipartUploadHandles) {
+        m_transferManager->AbortMultipartUpload(fileToHandle.second);
+      }
+    }
+    // remove temp folder if existing
+    auto tmpfolder = GetCacheTemporaryDirectory();
+    if (FileExists(tmpfolder, false) &&
+        IsDirectory(tmpfolder, true)) {         // log on
+      DeleteFilesInDirectory(tmpfolder, true);  // delete folder itself
+    }
+
+    m_client.reset();
+    m_transferManager.reset();
+    m_cache.reset();
+    m_directoryTree.reset();
+    m_unfinishedMultipartUploadHandles.clear();
+
+    m_cleanup.store(true);
+  }
 }
 
 // --------------------------------------------------------------------------
@@ -162,13 +171,16 @@ void Drive::SetDirectoryTree(unique_ptr<DirectoryTree> dirTree) {
 
 // --------------------------------------------------------------------------
 bool Drive::IsMountable() const {
-  m_mountable.store(Connect(false));  // synchronizely
+  m_mountable.store(Connect());
   return m_mountable.load();
 }
 
 // --------------------------------------------------------------------------
-bool Drive::Connect(bool buildupDirTreeAsync) const {
-  auto err = GetClient()->HeadBucket();
+bool Drive::Connect() const {
+  // Connect is call before fuse_main and thread pools are initialized in
+  // fuse init, so at the time the thread pool is still not been initialized.
+  bool notUseThreadPool = false;
+  auto err = GetClient()->HeadBucket(notUseThreadPool);
   if (!IsGoodQSError(err)) {
     DebugError(GetMessageForQSError(err));
     return false;
@@ -180,16 +192,11 @@ bool Drive::Connect(bool buildupDirTreeAsync) const {
   }
 
   // Build up the root level of directory tree asynchornizely.
-  auto ReceivedHandler = [](const ClientError<QSError> &err) {
+  auto DoListDirectory = [this, notUseThreadPool] {
+    auto err = GetClient()->ListDirectory("/", notUseThreadPool);
     DebugErrorIf(!IsGoodQSError(err), GetMessageForQSError(err));
   };
-
-  if (buildupDirTreeAsync) {  // asynchronizely
-    GetClient()->GetExecutor()->SubmitAsyncPrioritized(
-        ReceivedHandler, [this] { return GetClient()->ListDirectory("/"); });
-  } else {  // synchronizely
-    ReceivedHandler(GetClient()->ListDirectory("/"));
-  }
+  std::async(std::launch::async, DoListDirectory);
 
   return true;
 }
